requirements.txt
discord.py-self
python-dotenv
google-generativeai
ccxt
